fix(cron): rename plugins/cron to plugins/cron_providers to resolve WebUI import collision
